+++ HISTORY.rst
@@ -7,6 +7,14 @@ Pending Release
 
 .. Insert new release notes below this line
 
+* Converted setuptools metadata to configuration file. This meant removing the
+  ``__version__`` attribute from the package. If you want to inspect the
+  installed version, use
+  ``pkg_resources.get_distribution("flake8-tidy-imports").version``
+  (`docs <https://setuptools.readthedocs.io/en/latest/pkg_resources.html#getting-or-creating-distributions>`__).
+
+* Add dependencies on ``cached-property`` and ``setuptools``.
+
 3.0.0 (2019-10-15)
 ------------------